feat: separate page for signup and add components
