+++ client/src/lib/components/auth/OrDivider.svelte
@@ -1,0 +1,8 @@
+<div class="relative my-6">
+  <div class="absolute inset-0 flex items-center">
+    <div class="w-full border-t border-gray-300"></div>
+  </div>
+  <div class="relative flex justify-center text-sm">
+    <span class="px-2 bg-white text-xs text-gray-500">or</span>
+  </div>
+</div>
