+++ client/src/routes/auth/login/+page.svelte
@@ -1,5 +1,9 @@
 <script>
-  import { ChevronLeft, Eye, EyeOff } from 'lucide-svelte';
+  import AuthContainer from '$lib/components/auth/AuthContainer.svelte';
+  import AuthHeader from '$lib/components/auth/AuthHeader.svelte';
+  import PasswordInput from '$lib/components/auth/PasswordInput.svelte';
+  import GoogleOAuthButton from '$lib/components/auth/GoogleOAuthButton.svelte';
+  import OrDivider from '$lib/components/auth/OrDivider.svelte';
   
   let email = '';
   let password = '';
@@ -7,108 +11,60 @@
   let showPassword = false;
 
   function handleLogin() {
-    // Handle login logic here
     console.log('Login:', { email, password, rememberMe });
+  }
+  
+  function handleGoogleOAuth() {
+    console.log('Google OAuth login');
   }
 </script>
 
-<div class="min-h-screen flex items-center justify-center bg-gradient-to-br from-green-50 to-blue-50">
-  <div class="w-full max-w-lg p-10 bg-white rounded-sm border border-gray-200 shadow-sm">
-    <!-- Back arrow and title -->
-    <div class="flex items-center space-x-1 mb-6">
-      <ChevronLeft class="w-5 h-5 text-gray-600 cursor-pointer" on:click={() => window.history.back()} />
-      <h2 class="text-lg font-semibold text-black">Job Applier </h2>
-      <h2 class="text-lg font-semibold text-green-700">3000</h2>
+<AuthContainer>
+  <AuthHeader />
+  <h1 class="text-2xl font-semibold text-gray-900 mb-5">Sign in</h1>
+
+  <form on:submit|preventDefault={handleLogin} class="space-y-6">
+    <div>
+      <input
+        id="email"
+        type="email"
+        bind:value={email}
+        placeholder="Enter email..."
+        required
+        class="w-full px-4 py-3 text-sm border border-gray-300 rounded-md focus:ring-1 focus:ring-gray-500 outline-none transition-colors"
+      />
     </div>
-    <h1 class="text-2xl font-semibold text-gray-900 mb-5">Login to your account</h1>
 
+    <PasswordInput
+      id="password"
+      label=""
+      placeholder="Enter password..."
+      bind:value={password}
+      bind:showPassword
+    />
 
-    <!-- Login form -->
-    <form on:submit|preventDefault={handleLogin} class="space-y-6">
-      <!-- Email field -->
-      <div>
+    <div class="flex items-center justify-between">
+      <label class="flex items-center cursor-pointer">
         <input
-          id="email"
-          type="email"
-          bind:value={email}
-          placeholder="Enter email..."
-          required
-          class="w-full px-4 py-3 text-sm border border-gray-300 rounded-md focus:ring-1 focus:ring-gray-500 outline-none transition-colors"
+          type="checkbox"
+          bind:checked={rememberMe}
+          class="w-4 h-4 text-green-600 border-gray-300 rounded focus:ring-0 focus:ring-offset-0"
         />
-      </div>
-
-      <!-- Password field -->
-      <div>
-        <div class="relative">
-          <input
-            id="password"
-            type={showPassword ? 'text' : 'password'}
-            bind:value={password}
-            placeholder="Enter password..."
-            required
-            class="w-full px-4 py-3 text-sm pr-12 border border-gray-300 rounded-md focus:ring-1 focus:ring-gray-500 outline-none transition-colors"
-          />
-          <button
-            type="button"
-            on:click={() => showPassword = !showPassword}
-            class="absolute right-3 top-1/2 -translate-y-1/2 text-gray-500 hover:text-gray-700"
-            aria-label={showPassword ? 'Hide password' : 'Show password'}
-          >
-            {#if showPassword}
-              <EyeOff class="w-4 h-4" />
-            {:else}
-              <Eye class="w-4 h-4" />
-            {/if}
-          </button>
-        </div>
-      </div>
-
-      <!-- Remember me and Forgot password -->
-      <div class="flex items-center justify-between">
-        <label class="flex items-center cursor-pointer">
-          <input
-            type="checkbox"
-            bind:checked={rememberMe}
-            class="w-4 h-4 text-green-600 border-gray-300 rounded focus:ring-0 focus:ring-offset-0"
-          />
-          <span class="ml-2 text-sm text-gray-700">Remember me</span>
-        </label>
-        <a href="/auth/forgot-password" class="text-sm text-green-600 hover:text-green-700 font-medium">
-          Forgot password
-        </a>
-      </div>
-
-      <!-- Login button -->
-      <button
-        type="submit"
-        class="w-full py-2 px-4 bg-green-600 hover:bg-green-700 text-white text-sm font-medium rounded-md transition-colors focus:outline-none focus:ring-1 focus:ring-green-500"
-      >
-        Log in
-      </button>
-    </form>
-
-    <!-- Divider -->
-    <div class="relative my-6">
-      <div class="absolute inset-0 flex items-center">
-        <div class="w-full border-t border-gray-300"></div>
-      </div>
-      <div class="relative flex justify-center text-sm">
-        <span class="px-2 bg-white text-gray-500">or</span>
-      </div>
+        <span class="ml-2 text-sm text-gray-700">Remember me</span>
+      </label>
+      <a href="/auth/forgot-password" class="text-sm text-green-600 hover:text-green-700 font-medium">
+        Forgot password
+      </a>
     </div>
 
-    <!-- Google OAuth button -->
     <button
-      type="button"
-      class="w-full flex items-center justify-center gap-3 py-2 px-4 border border-gray-300 rounded-md hover:bg-gray-50 transition-colors focus:outline-none focus:ring-1 focus:ring-gray-500"
+      type="submit"
+      class="w-full py-2 px-4 bg-green-600 hover:bg-green-700 text-white text-sm font-medium rounded-md transition-colors focus:outline-none focus:ring-1 focus:ring-green-500"
     >
-      <svg class="w-5 h-5" viewBox="0 0 24 24">
-        <path fill="#4285F4" d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"/>
-        <path fill="#34A853" d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"/>
-        <path fill="#FBBC05" d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"/>
-        <path fill="#EA4335" d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"/>
-      </svg>
-      <span class="text-sm font-medium text-gray-700">Continue with Google</span>
+      Log in
     </button>
-  </div>
-</div>
+  </form>
+
+  <OrDivider />
+  <GoogleOAuthButton onClick={handleGoogleOAuth} />
+</AuthContainer>
